Refactor statistics logic into manager classes
Moved chart and telemetry logic from StatisticsController to new StatisticsChartManager and TelemetryManager classes for improved separation of concerns and maintainability. StatisticsController now delegates chart setup and telemetry updates to these managers.
