+++ .idea/workspace.xml
@@ -5,8 +5,8 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="940e75ed-a3fe-4fda-aab8-c6d80cc05012" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/controllers/ParliamentController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/controllers/ParliamentController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/controllers/StartController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/controllers/StartController.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/managers/StatisticsChartManager.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/de/schulprojekt/duv/view/managers/TelemetryManager.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
